+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Change log
+
+### 1.5.0
+#### Breaking changes
+- `reduceRightUntil` changed to `reduceUntilRight`.
+- `lengthsToSmallest` changed to `listsToSmallest`.
+
 ### 1.3.0
 - Added `noop` (no-op (op as in operation)) method (useful as a placeholder for variables/properties that should always contain a function).
 - Added test for `noop` addition.